+++ examples/vite.config.local.mjs
@@ -37,9 +37,6 @@ export default defineConfig(async () => {
     server: {
       open: true,
       port: 8080
-    },
-    optimizeDeps: {
-      esbuildOptions: {target: 'es2022'}
     }
   };
 });